# Doctors/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"RegisterUpdateResDoctor": {
				"properties": {
					"like": {},
					"_id": {
						"type": "string"
					},
					"FirstName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"like",
					"_id",
					"FirstName",
					"email",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"loginDoctor": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegisterUpdateResPatient": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"FullName": {
						"type": "string"
					},
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExpires": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "number",
						"format": "double"
					},
					"otpCode": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"FullName",
					"resetPasswordToken",
					"resetPasswordExpires",
					"email",
					"otpCode",
					"cell",
					"password",
					"adress",
					"disease",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterPatient": {
				"properties": {
					"FullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					},
					"disease": {
						"type": "string"
					},
					"doctor": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"FullName",
					"email",
					"cell",
					"password",
					"adress",
					"disease",
					"doctor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"loginPatient": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPatient": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailVerifyPatient": {
				"properties": {
					"resetPasswordToken": {
						"type": "string"
					},
					"resetPasswordExpires": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "number",
						"format": "double"
					},
					"otpCode": {
						"type": "string"
					}
				},
				"required": [
					"resetPasswordToken",
					"resetPasswordExpires",
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"forgotPassword": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VerifyOtpCodes": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otpCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"verifyOTPCode": {
				"properties": {
					"email": {
						"type": "string"
					},
					"otpCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"email",
					"otpCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"changePassword": {
				"properties": {
					"email": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"newPassword",
					"confirmPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"searchPatient": {
				"properties": {
					"FullName": {
						"type": "string"
					}
				},
				"required": [
					"FullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"searchPatientByLocation": {
				"properties": {
					"disease": {
						"type": "string"
					},
					"adress": {
						"type": "string"
					}
				},
				"required": [
					"disease",
					"adress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeletePatient": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "admin_token",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "Bearer",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swaggerapis2",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/doctor/loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResDoctor"
								}
							}
						}
					}
				},
				"tags": [
					"doctor"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginDoctor"
							}
						}
					}
				}
			}
		},
...
		"/patient/patient-register": {
			"post": {
				"operationId": "RegisterPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterPatient"
							}
						}
					}
				}
			}
		},
		"/patient/patient-loginUser": {
			"post": {
				"operationId": "LoginEmailUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/loginPatient"
							}
						}
					}
				}
			}
		},
...
		"/patient/getPatient": {
			"post": {
				"operationId": "GetPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetPatient"
							}
						}
					}
				}
			}
		},
...
		"/patient/forgot-password": {
			"post": {
				"operationId": "ForgotPass",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmailVerifyPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/forgotPassword"
							}
						}
					}
				}
			}
		},
		"/patient/verify-OTPCode": {
			"post": {
				"operationId": "VerifyEamil",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VerifyOtpCodes"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/verifyOTPCode"
							}
						}
					}
				}
			}
		},
		"/patient/change-password": {
			"post": {
				"operationId": "ChangePasswords",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/changePassword"
							}
						}
					}
				}
			}
		},
		"/patient/search": {
			"post": {
				"operationId": "SearchPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/searchPatient"
							}
						}
					}
				}
			}
		},
		"/patient/search-by-location": {
			"post": {
				"operationId": "SearchByLocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/searchPatientByLocation"
							}
						}
					}
				}
			}
		},
		"/patient/delete": {
			"delete": {
				"operationId": "DeletePatients",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RegisterUpdateResPatient"
								}
							}
						}
					}
				},
				"tags": [
					"patient"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletePatient"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}